Add files via upload
Project Description:
This research project aims to understand the influence of social media on the Ukraine-Russia war. We collaborated with the Taiwan Digital Diplomacy Association and Meta to collect data from 200,000+ posts on Facebook, Telegram, and Twitter from officials like Ukraine President “Volodymyr Zelenskyy,” the Ukrainian ministry of foreign affairs, and local newspapers. We analyzed public response, sentimental behavior, and trend creation on social media platforms through topic modeling and sentiment analysis in Python.
The results showed that most tweets expressed condolences and support, with the highest tweet density in the US, Ukraine, and Iran. We formulated a guidance proposal for the government of Taiwan for effective digital communication with Taiwanese citizens during times of conflict through increased awareness of public sentiments.